LayoutTests/fast/js/kde/parse-expected.txt

KDE JS Test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS res is 11
PASS caught is true
PASS successfullyParsed is true

TEST COMPLETE
